After load set webview to opaque
During the initial load we need the clear + !opaque settings so
the webview does not flash a white background in dark mode. But
after load we want to use the systemBackground colour and an
opaque background, so in dark mode the scrollbars are displayed
correctly. Without this iOS will display a dark scrollbar ontop
of the dark background.

Note that due to a bug in Webkit setting the scrollbar inset
colours doesn't seem to work in current iOS:

https://developer.apple.com/forums/thread/689654