Reset tabIndex when state is reset to loading
We track the state of the current index in two places: the audio
controller, which updates as audio as played, the tabView has its
own binding of the current index, this is to allow scrolling.

When the audio item being played changes, we need to reset the
state of the tabIndex. We do this by watching for state loading
which occurs when new audio is loaded.