Dont crash if clicking on a fragment link
If a user clicks on a fragment link like <a href="#info">click here to skip to info</a>
we would previously crash because we would reload the bundle.

This prevents the crash but the link does not work, we can't
just synthesize these links with JS, because the id="info" items
are usually removed by readability.

To properly handle this, we probably need to preserve some id
elements if they are linked to. Then we can scroll to those elements
in Javascript.