Fix hover behavior of library list hover actions
In the library list view, there is a toolbar of actions that appears over each
entry when you hover over it. This change fixes the hover behavior of this
toolbar so that you do not need to move the mouse cursor out and back into the
library list item after archiving/deleting a previous item in order to
re-activate the toolbar.