Create a debug package name
# Context

This will allow developers to have both production and dev version on their phones at the same time. Also changed the name of the Dev version to Dev - Omnivore

If someone has the orginal icon files, I would prefer to change the debug icon to be a different color to distinguish it more. But  I couldn't find the main psd or svg file that the pngs are generated from