fix: page becomes unresponsive after closing modal
Came across this issue while migrating to Omnivore, but I saw it also
reported in issue #3648. There was a first attempt to fix in #3653, but
I believe it's an issue in the library itself, which is not there
anymore in 2.0.6.

It also fixes the "too much recursion" error that shows in the console
when the modal is open, both in the Feeds page and in the Inbox (e.g.
when clicking on Edit Info).